Move some helper functions into a helper mod
